requirements.txt
github3.py==0.9.6
GitPython==2.1.7
requests==2.18.4
neo4j-driver==1.5.0
python-gitlab==1.2.0
lxml==4.1.1
html5lib==1.0.1
cssselect==1.0.3
